backend-services/event-service/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.27.0
sqlalchemy==2.1.0
python-dotenv==1.0.0
pydantic==1.10.12
requests==2.31.0
requests==2.31.0
